requirements.txt
python-bitcoinrpc>=1.0
redis==3.5.2
mysql-connector-python>=8.0.16
